file_merger_trending_topics.py
I wrote a scraper that scrapes the trending topics in India every day. I used Python Anywhere and it scraped about 6 months of data as a .csv. I'm working on refactoring the code to write the data to a database instead of a .csv so I need to take the existing data and combine it and clean it and ingest it into the database. This script's purpose is to open the existing .csvs from a folder on my local device, ingest all the data into a dataframe, clean it and sort it, and export it back out as a larger .csv.